+++ source/isaaclab_tasks/test/benchmarking/test_environments_training.py
@@ -18,8 +18,8 @@ import sys
 import time
 
 import carb
+import env_benchmark_test_utils as utils
 import pytest
-import test_utils as utils
 
 from isaaclab.utils.pretrained_checkpoint import WORKFLOW_EXPERIMENT_NAME_VARIABLE, WORKFLOW_TRAINER